% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSMDplot.R
\name{createSMDplot}
\alias{createSMDplot}
\title{Create a plot of the standardized mean differences in covariates}
\usage{
createSMDplot(matched.model)
}
\arguments{
\item{matched.model}{A matchitobject returned from the matchit function. \code{matched.model}}
}
\value{
NA
}
\description{
This function takes a matchitobject and returns a plot of the standardized
mean differences (SMD) between the high exposed and control locations for
each covariate in the propensity score model.  The SMD is a common way to
evalulate whether covariates were balanced between the two groups during matching.
Code for this plot was adopted from a vignette to the R tableone package (Yoshida and Bohn).
}
\examples{
# Include these regions
regions <- c("IndustrialMidwest", "Northeast", "Southeast")

# Covariates to adjust for using propensity score matching
covariate.vars <- c("logPop", "PctUrban","MedianHHInc", "PctPoor", "smokerate2000")

dataset <- getMatchedDataset(exposure = inmap2005, covariates, covariate.vars, regions)

createSMDplot(dataset$matched.model)
}
\references{
\insertRef{austin2011introduction}{sourceOrientedApproach}

\insertRef{yoshidapackage}{sourceOrientedApproach}

\insertRef{ho2011matchit}{sourceOrientedApproach}
}
\keyword{keywords}
