% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOutcomeModel.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Fit a Poisson regression model to the health outcome data}
\usage{
fitOutcomeModel(dataset, outcome, covariate.vars, outcome.var = "IHD",
  offset.var = "person_years", analysis.type = "matched")
}
\arguments{
\item{dataset}{A data table.  If a matched analysis is performed, then this should be
the matched dataset.  If a stratified analysis is performed, this should be the raw
dataset. \code{dataset}}

\item{outcome}{A data table containing the outcome data.  The first column should be
the ZIP code.\code{outcome}}

\item{covariate.vars}{A character vector containing the names of covariates in the dataset
to adjust for in the outcome model. Must be a subset of the colnames of dataset. \code{covariate.vars}}

\item{outcome.var}{The name of the outcome variable in the outcome data table. \code{outcome.var}}

\item{offset.var}{The name of the variable in the outcome data table to use as an offset in the
Poisson regression model.  Typically, this is the number of person years observed. \code{offset.var}}

\item{analysis.type}{The type of analysis to perform.  Either "matched" or "stratified".  A
"matched" analysis should use the matched dataset.  A "stratified" analysis should use the raw
dataset. \code{analysis.type}}
}
\value{
An object of the glm class.
}
\description{
This function takes a dataset from the getMatchedDataset function and fits
a Poisson regression model.  There are two types of analyses that can be performed.
First, using the matched dataset, a matched analysis can be performed.  Second,
using the raw dataset, a stratified analysis can be performed. log(offset.var) is
used as an offset in the model.
}
\examples{
data('ihd2005')
data('inmap2005')
data('covariates')

# Include these regions
regions <- c("IndustrialMidwest", "Northeast", "Southeast")

# Covariates to adjust for using propensity score matching
covariate.vars <- c("logPop", "PctUrban","MedianHHInc", "PctPoor", "smokerate2000")

dataset <- getMatchedDataset(exposure = inmap2005, covariates, covariate.vars, regions)

fitOutcomeModel(dataset$matched, ihd2005, covariate.vars)
}
\keyword{keywords}
