% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchedDataset.R
\name{getMatchedDataset}
\alias{getMatchedDataset}
\title{Obtain a propensity score matched dataset}
\usage{
getMatchedDataset(exposure, covariates, covariate.vars, regions = "all",
  exact.vars = NULL, exposure.cutoff.percentile = 0.8,
  caliper = "default")
}
\arguments{
\item{exposure}{A data table.  The first column must be the unit identifier (for example, 5-digit U.S. zip code)
and the second column is the exposure of interest \code{exposure}}

\item{covariates}{A data table. The first column must be the unit identifier (for example, 5-digit U.S. zip code).
Subsequent columns are covariates to include in the propensity score model. \code{covariates}}

\item{covariate.vars}{A character vector containing a subset of
the covariate column names to include in the propensity score model. \code{covariate.vars}}

\item{regions}{A character vector containing the geographic regions
("Northeast", "Southeast", etc.) to include in the analysis. Defaults to
all regions. \code{regions}}

\item{exact.vars}{A character vector of colnames in covariates to match exactly
on.  This variable cannot be a continuous variable.\code{exact.vars}}

\item{exposure.cutoff.percentile}{A numeric value between 0 and 1 specifying the
cutoff, in terms of percentile of the exposure distribution, between high exposed
and controls locations.}

\item{caliper}{A numeric value specifying the maximum allowable difference
in propensity scores when performing matching.}
}
\value{
output A list containing the raw dataset, the matched dataset,
a matchitobject, the caliper, and the high/control cutoff.
}
\description{
This function is designed for use with the analysis performed in
A Source Oriented Approach to Coal Emissions Health Effects.  Specifically,
this function dichotomizes the exposure variable (typically, coal emissions)
to classify U.S. ZIP codes
as either highly exposed or control locations. Using the MatchIt package, high exposed
ZIP codes are matched to control locations with similar propensity scores.
A caliper can be specified, which is the maximum allowed difference in
propensity scores when
selecting matches.  The default caliper is 20 percent of the pooled standard
deviation of the logit of the propensity scores of the high exposed and
control locations, as suggested in Austin 2011.  Categorical variables
can be matched on exactly.
}
\examples{
# Include these regions
regions <- c("IndustrialMidwest", "Northeast", "Southeast")

# Covariates to adjust for using propensity score matching
covariate.vars <- c("logPop", "PctUrban","MedianHHInc", "PctPoor", "smokerate2000")

dataset <- getMatchedDataset(exposure = inmap2005, covariates, covariate.vars, regions)
}
\references{
\insertRef{austin2011introduction}{sourceOrientedApproach}

\insertRef{ho2011matchit}{sourceOrientedApproach}
}
\keyword{keywords}
