% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZIPcodeCovariates.R
\name{getZIPcodeCovariates}
\alias{getZIPcodeCovariates}
\title{Get US ZIP Code Covariates}
\usage{
getZIPcodeCovariates()
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
Return a data table consisting of CENSUS 2000 variables, smoking rates,
and weather variables for every US ZIP code
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
