% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZIPcodeOutcomes.R
\name{getZIPcodeOutcomes}
\alias{getZIPcodeOutcomes}
\title{A one sentence description of what your function does}
\usage{
getZIPcodeOutcomes(year = 2005)
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
