% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExposureUSmap.R
\name{plotExposureUSmap}
\alias{plotExposureUSmap}
\title{Plot a U.S. map depicting the exposure variable}
\usage{
plotExposureUSmap(dataset, exposure.binary = TRUE, main = "")
}
\arguments{
\item{dataset}{A data table containing the exposure variable.  If continuous, the exposure
variable must be the second column of the data table.  If binary, the exposure should be
either one or zero and in a column named High. In addition, the data table must contain
each ZIP code's coordinates in columns named Longtitude and Latitude \code{dataset}}

\item{exposure.binary}{TRUE/FALSE. Whether to plot the binary or continuous exposure.
\code{exposure.binary}}
}
\value{
NA
}
\description{
This function plots a map depicting the exposure variable on a U.S. map.
The exposure variable can be either continous or binary.
}
\examples{
data('covariates')
data('inmap2005')
covariate.vars <- c("logPop", "PctUrban", "PctBlack","MedianHHInc","smokerate2000")
dataset <- getMatchedDataset(exposure, covariates, covariate.vars)

#plot the raw dataset
plotExposureUSmap(dataset$raw, exposure.binary = TRUE)
plotExposureUSmap(dataset$raw, exposure.binary = FALSE)
}
\references{
\insertRef{kahle2013ggmap}{sourceOrientedApproach}
}
\keyword{keywords}
