% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPropensityScoreHistogram.R
\name{plotPropensityScoreHistogram}
\alias{plotPropensityScoreHistogram}
\title{Plot a histogram of the propensity score by exposure group}
\usage{
plotPropensityScoreHistogram(dataset, main = "")
}
\arguments{
\item{dataset}{A data table.  Two of the column names must be High and prop.scores. \code{dataset}}

\item{main}{Title for the plot. \code{main}}
}
\value{
NA
}
\description{
This function plots a histogram of the propensity scores for the high exposed
and control locations.
}
\examples{
# Include these regions
regions <- c("IndustrialMidwest", "Northeast", "Southeast")

# Covariates to adjust for using propensity score matching
covariate.vars <- c("logPop", "PctUrban","MedianHHInc", "PctPoor", "smokerate2000")

dataset <- getMatchedDataset(exposure = inmap2005, covariates, covariate.vars, regions)

plotPropensityScoreHistogram(dataset$raw)
}
\keyword{keywords}
